backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.30
pydantic>=2.10.0
python-dotenv>=1.0.0
yfinance>=0.2.40
httpx>=0.27.0
